Moved conversion functions into their own module.  Refactored and removed duplicate functions.
