+++ tests/tests.py
@@ -1,5 +1,8 @@
 from gunfolds.tools import bfutils
+from gunfolds.tools import conversions
 from gunfolds.tools import unknownrate as ur
+from gunfolds.tools import zickle as zkl
+import os
 import unittest
 
 
@@ -7,9 +10,8 @@ class TestBFUtilsFunctions(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        # Read in pickle file with results
-        G = {'1': {'1': set([(0, 1)]),
-                   '2': set([(0, 1)]), 
+        cls._G = {'1': {'1': set([(0, 1)]),
+                   '2': set([(0, 1)]),
                    '3': set([(0, 1)])},
              '3': {'1': set([(0, 1)]),
                    '3': set([(0, 1)]),
@@ -24,13 +26,13 @@ class TestBFUtilsFunctions(unittest.TestCase):
                    '2': set([(0, 1)]),
                    '5': set([(0, 1)])}}
 
-    def test_call_undersamples(self):
+    def test__call_undersamples(self):
         gs = [{'1': {'1': {(0, 1)}, '2': {(0, 1)}, '3': {(0, 1)}},
               '2': {'2': {(0, 1)}, '3': {(0, 1)}, '4': {(0, 1)}},
               '3': {'1': {(0, 1)}, '2': {(0, 1)}, '3': {(0, 1)}, '4': {(0, 1)}},
               '4': {'1': {(0, 1)}, '2': {(0, 1)}, '3': {(0, 1)}, '5': {(0, 1)}},
               '5': {'1': {(0, 1)}}},
-             {'1': {'1': {(0, 1)},
+              {'1': {'1': {(0, 1)},
                '2': {(0, 1), (2, 0)},
                '3': {(0, 1), (2, 0)},
                '4': {(0, 1), (2, 0)},
@@ -50,7 +52,7 @@ class TestBFUtilsFunctions(unittest.TestCase):
                '3': {(0, 1), (2, 0)},
                '4': {(0, 1)}},
               '5': {'1': {(0, 1), (2, 0)}, '2': {(0, 1), (2, 0)}, '3': {(0, 1), (2, 0)}}},
-                    {'1': {'1': {(0, 1)},
+              {'1': {'1': {(0, 1)},
                '2': {(0, 1), (2, 0)},
                '3': {(0, 1), (2, 0)},
                '4': {(0, 1), (2, 0)},
@@ -74,7 +76,7 @@ class TestBFUtilsFunctions(unittest.TestCase):
                '2': {(0, 1), (2, 0)},
                '3': {(0, 1), (2, 0)},
                '4': {(0, 1), (2, 0)}}},
-             {'1': {'1': {(0, 1)},
+              {'1': {'1': {(0, 1)},
                '2': {(0, 1), (2, 0)},
                '3': {(0, 1), (2, 0)},
                '4': {(0, 1), (2, 0)},
@@ -100,67 +102,162 @@ class TestBFUtilsFunctions(unittest.TestCase):
                '4': {(0, 1), (2, 0)},
                '5': {(0, 1)}}}]
 
-        gs_test = bfutils.call_undersamples(self.G)
+        gs_test = bfutils.call_undersamples(self._G)
         self.assertEqual(gs, gs_test)
 
+    def test__call_undersample(self):
+        u = 1
+        g_u_1 = {'1': {'1': {(0, 1)},
+                       '2': {(0, 1), (2, 0)},
+                 '3': {(0, 1), (2, 0)},
+                       '4': {(0, 1), (2, 0)},
+                       '5': {(2, 0)}},
+                 '2': {'1': {(0, 1), (2, 0)},
+                       '2': {(0, 1)},
+                       '3': {(0, 1), (2, 0)},
+                       '4': {(0, 1), (2, 0)},
+                       '5': {(0, 1), (2, 0)}},
+                 '3': {'1': {(0, 1), (2, 0)},
+                       '2': {(0, 1), (2, 0)},
+                       '3': {(0, 1)},
+                       '4': {(0, 1), (2, 0)},
+                       '5': {(0, 1), (2, 0)}},
+                 '4': {'1': {(0, 1), (2, 0)},
+                       '2': {(0, 1), (2, 0)},
+                       '3': {(0, 1), (2, 0)},
+                       '4': {(0, 1)}},
+                 '5': {'1': {(0, 1), (2, 0)}, '2': {(0, 1), (2, 0)}, '3': {(0, 1), (2, 0)}}}
+        g2 = bfutils.undersample(self._G, u)
+        self.assertEqual(g_u_1, g2)
 
-    def test_call_undersample(self):
         u = 2
         g_u_2 = {'1': {'1': {(0, 1)},
-              '2': {(0, 1), (2, 0)},
-              '3': {(0, 1), (2, 0)},
-              '4': {(0, 1), (2, 0)},
-              '5': {(0, 1), (2, 0)}},
-             '2': {'1': {(0, 1), (2, 0)},
-              '2': {(0, 1)},
-              '3': {(0, 1), (2, 0)},
-              '4': {(0, 1), (2, 0)},
-              '5': {(0, 1), (2, 0)}},
-             '3': {'1': {(0, 1), (2, 0)},
-              '2': {(0, 1), (2, 0)},
-              '3': {(0, 1)},
-              '4': {(0, 1), (2, 0)},
-              '5': {(0, 1), (2, 0)}},
-             '4': {'1': {(0, 1), (2, 0)},
-              '2': {(0, 1), (2, 0)},
-              '3': {(0, 1), (2, 0)},
-              '4': {(0, 1)},
-              '5': {(0, 1), (2, 0)}},
-             '5': {'1': {(0, 1), (2, 0)},
-              '2': {(0, 1), (2, 0)},
-              '3': {(0, 1), (2, 0)},
-              '4': {(0, 1), (2, 0)}}}
-        g2 = bfutils.undersample(self.G,u)
+                       '2': {(0, 1), (2, 0)},
+                 '3': {(0, 1), (2, 0)},
+                       '4': {(0, 1), (2, 0)},
+                       '5': {(0, 1), (2, 0)}},
+                 '2': {'1': {(0, 1), (2, 0)},
+                       '2': {(0, 1)},
+                       '3': {(0, 1), (2, 0)},
+                       '4': {(0, 1), (2, 0)},
+                       '5': {(0, 1), (2, 0)}},
+                 '3': {'1': {(0, 1), (2, 0)},
+                       '2': {(0, 1), (2, 0)},
+                       '3': {(0, 1)},
+                       '4': {(0, 1), (2, 0)},
+                       '5': {(0, 1), (2, 0)}},
+                 '4': {'1': {(0, 1), (2, 0)},
+                       '2': {(0, 1), (2, 0)},
+                       '3': {(0, 1), (2, 0)},
+                       '4': {(0, 1)},
+                       '5': {(0, 1), (2, 0)}},
+                 '5': {'1': {(0, 1), (2, 0)},
+                       '2': {(0, 1), (2, 0)},
+                       '3': {(0, 1), (2, 0)},
+                       '4': {(0, 1), (2, 0)}}}
+        g2 = bfutils.undersample(self._G, u)
         self.assertEqual(g_u_2, g2)
 
         u = 4
         g_u_4 = {'1': {'1': {(0, 1)},
-              '2': {(0, 1), (2, 0)},
-              '3': {(0, 1), (2, 0)},
-              '4': {(0, 1), (2, 0)},
-              '5': {(0, 1), (2, 0)}},
-             '2': {'1': {(0, 1), (2, 0)},
-              '2': {(0, 1)},
-              '3': {(0, 1), (2, 0)},
-              '4': {(0, 1), (2, 0)},
-              '5': {(0, 1), (2, 0)}},
-             '3': {'1': {(0, 1), (2, 0)},
-              '2': {(0, 1), (2, 0)},
-              '3': {(0, 1)},
-              '4': {(0, 1), (2, 0)},
-              '5': {(0, 1), (2, 0)}},
-             '4': {'1': {(0, 1), (2, 0)},
-              '2': {(0, 1), (2, 0)},
-              '3': {(0, 1), (2, 0)},
-              '4': {(0, 1)},
-              '5': {(0, 1), (2, 0)}},
-             '5': {'1': {(0, 1), (2, 0)},
-              '2': {(0, 1), (2, 0)},
-              '3': {(0, 1), (2, 0)},
-              '4': {(0, 1), (2, 0)},
-              '5': {(0, 1)}}}
-        g2 = bfutils.undersample(self.G,u)
+                       '2': {(0, 1), (2, 0)},
+                 '3': {(0, 1), (2, 0)},
+                       '4': {(0, 1), (2, 0)},
+                       '5': {(0, 1), (2, 0)}},
+                 '2': {'1': {(0, 1), (2, 0)},
+                       '2': {(0, 1)},
+                       '3': {(0, 1), (2, 0)},
+                       '4': {(0, 1), (2, 0)},
+                       '5': {(0, 1), (2, 0)}},
+                 '3': {'1': {(0, 1), (2, 0)},
+                       '2': {(0, 1), (2, 0)},
+                       '3': {(0, 1)},
+                       '4': {(0, 1), (2, 0)},
+                       '5': {(0, 1), (2, 0)}},
+                 '4': {'1': {(0, 1), (2, 0)},
+                       '2': {(0, 1), (2, 0)},
+                       '3': {(0, 1), (2, 0)},
+                       '4': {(0, 1)},
+                       '5': {(0, 1), (2, 0)}},
+                 '5': {'1': {(0, 1), (2, 0)},
+                       '2': {(0, 1), (2, 0)},
+                       '3': {(0, 1), (2, 0)},
+                       '4': {(0, 1), (2, 0)},
+                       '5': {(0, 1)}}}
+        g2 = bfutils.undersample(self._G, u)
         self.assertEqual(g_u_4, g2)
+
+
+
+
+class TestConversionFunctions(unittest.TestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        cls._G = {'1': {'1': set([(0, 1)]),
+                   '2': set([(0, 1)]),
+                   '3': set([(0, 1)])},
+             '3': {'1': set([(0, 1)]),
+                   '3': set([(0, 1)]),
+                   '2': set([(0, 1)]),
+                   '4': set([(0, 1)])},
+             '2': {'3': set([(0, 1)]),
+                   '2': set([(0, 1)]),
+                   '4': set([(0, 1)])},
+             '5': {'1': set([(0, 1)])},
+             '4': {'1': set([(0, 1)]),
+                   '3': set([(0, 1)]),
+                   '2': set([(0, 1)]),
+                   '5': set([(0, 1)])}}
+
+
+    def test__dict_format_converter(self):
+        expected = {1: {1: 1, 2: 1, 3: 1},
+                    2: {2: 1, 3: 1, 4: 1},
+                    3: {1: 1, 2: 1, 3: 1, 4: 1},
+                    4: {1: 1, 2: 1, 3: 1, 5: 1},
+                    5: {1: 1}}
+        converted = conversions.dict_format_converter(self._G)
+        self.assertEqual(expected, converted)
+
+
+
+
+class TestUnknownRateFunctions(unittest.TestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        # Read in pickle file with results
+        cls._G = {'1': {'1': set([(0, 1)]),
+                   '2': set([(0, 1)]),
+                   '3': set([(0, 1)])},
+             '3': {'1': set([(0, 1)]),
+                   '3': set([(0, 1)]),
+                   '2': set([(0, 1)]),
+                   '4': set([(0, 1)])},
+             '2': {'3': set([(0, 1)]),
+                   '2': set([(0, 1)]),
+                   '4': set([(0, 1)])},
+             '5': {'1': set([(0, 1)])},
+             '4': {'1': set([(0, 1)]),
+                   '3': set([(0, 1)]),
+                   '2': set([(0, 1)]),
+                   '5': set([(0, 1)])}}
+        DIR_NAME = os.path.dirname(__file__)
+        cls._ABS_PATH = os.path.abspath(os.path.join(DIR_NAME))
+
+    def test__liteqclass(self):
+        u = 1
+        set_of_u_1 = zkl.load("{}/liteqclass_output_n_5_u_1.zkl".format(self._ABS_PATH))
+        g2 = bfutils.undersample(self._G, u)
+        s = ur.liteqclass(g2, verbose=False, capsize=1000)
+        self.assertEqual(s, set_of_u_1)
+
+        u = 2
+        set_of_u_2 = zkl.load("{}/liteqclass_output_n_5_u_2.zkl".format(self._ABS_PATH))
+        g2 = bfutils.undersample(self._G, u)
+        s = ur.liteqclass(g2, verbose=False, capsize=1000)
+        self.assertEqual(s, set_of_u_2)
 
 
 if __name__ == '__main__':
